move handler chain implementation in its own file
